LayoutTests/inspector/debugger/command-line-api-exception-expected.txt

CONSOLE MESSAGE: line 4: TypeError: undefined is not an object (evaluating '[].x.x')
CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: variableThatDoesNotExist
CONSOLE MESSAGE: line 16: SyntaxError: Unexpected token ')'
CONSOLE MESSAGE: line 22: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
CONSOLE MESSAGE: line 27: thrown string
CONSOLE MESSAGE: line 32: 123.456
CONSOLE MESSAGE: line 37: null
CONSOLE MESSAGE: line 42: [object Object]
CONSOLE MESSAGE: line 47: [object HTMLBodyElement]
CONSOLE MESSAGE: line 55: CATCHER: TypeError: undefined is not an object (evaluating '[].x.x')
CONSOLE MESSAGE: line 55: CATCHER: thrown string
CONSOLE MESSAGE: line 55: CATCHER: [object Object]
Checks that $exception is available and accurate in evaluations when paused on an exception.


== Running test suite: CommandLineAPI.$exception
-- Running test case: BeforeExceptions
PASS: $exception should be undefined if there is no exception.

-- Running test case: UncaughtTypeException
$exception => TypeError: undefined is not an object (evaluating '[].x.x')

-- Running test case: UncaughtReferenceException
$exception => ReferenceError: Can't find variable: variableThatDoesNotExist

-- Running test case: UncaughtSyntaxException
$exception => SyntaxError: Unexpected token ')'

-- Running test case: UncaughtDOMException
$exception => IndexSizeError (DOM Exception 1): The index is not in the allowed range.

-- Running test case: UncaughtString
$exception => thrown string

-- Running test case: UncaughtNumber
$exception => 123.456

-- Running test case: UncaughtNull
$exception => null

-- Running test case: UncaughtObject
$exception => Object

-- Running test case: UncaughtNode
$exception => <body>

-- Running test case: CatchTypeException
PASS: Paused, stepping out to catch block...
PASS: `$exception` should be equal to `e`.
$exception => TypeError: undefined is not an object (evaluating '[].x.x')

-- Running test case: CatchThrownString
PASS: Paused, stepping out to catch block...
PASS: `$exception` should be equal to `e`.
$exception => thrown string

-- Running test case: CatchThrownObject
PASS: Paused, stepping out to catch block...
PASS: `$exception` should be equal to `e`.
$exception => Object

-- Running test case: AfterExceptions
PASS: $exception should be undefined if there is no exception.
